Extract validation logic into JobApplicationFile::Validatable concern
