+++ app/models/job_application_file.rb
@@ -3,6 +3,8 @@
 # Container for real file mandatory to fullfill a job application
 class JobApplicationFile < ApplicationRecord
   include Securable
+  include Validatable
+
   attr_accessor :job_application_file_existing_id, :do_not_provide_immediately
 
   belongs_to :job_application
@@ -31,36 +33,6 @@ class JobApplicationFile < ApplicationRecord
     else
       false
     end
-  end
-
-  def check!(administrator)
-    unless job_application_file_type.can_validate?(administrator)
-      errors.add(:base, :not_authorized_to_validate)
-      return false
-    end
-
-    update_column :is_validated, 1 # rubocop:disable Rails/SkipsModelValidations
-  end
-
-  def uncheck!(administrator)
-    unless job_application_file_type.can_validate?(administrator)
-      errors.add(:base, :not_authorized_to_validate)
-      return false
-    end
-
-    update_column :is_validated, 2 # rubocop:disable Rails/SkipsModelValidations
-  end
-
-  def validated?
-    is_validated == 1
-  end
-
-  def rejected?
-    is_validated == 2
-  end
-
-  def waiting_validation?
-    is_validated == 0
   end
 
   def downloadable? = JobApplication.states[job_application_state] < max_downloadable_state
